# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/modules": {
            "get": {
                "tags": [
                    "Modulos"
                ],
                "summary": "Mostrar modulos",
                "operationId": "2e94a3c5aa6628f36f22988ceb189fcc",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los modulos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Modulos"
                ],
                "summary": "Create a module",
                "operationId": "ba4fe3a1c551c9128c95e18e680e2bc7",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The title of the module",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Los campos no son validos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/modules/{id}": {
            "get": {
                "tags": [
                    "Modulos"
                ],
                "summary": "Mostrar un modulo concreto",
                "operationId": "a37173ec33ec4f217431701c5d603d73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar el modulo especificado."
                    },
                    "default": {
                        "description": "Error al procesar la solicitud"
                    }
                }
            },
            "put": {
                "tags": [
                    "Modulos"
                ],
                "summary": "Update a modules",
                "operationId": "25c4e23296978bc56610e638d49a37b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The title of the modules",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al procesar la solicitud', 'message"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Modulos"
                ],
                "summary": "Eliminar un modulo concreto",
                "operationId": "7362cedaa8c4ee544a48c9288a6819ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Aceptado."
                    },
                    "204": {
                        "description": "No"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Mostrar Roles",
                "operationId": "6cd48af64dafcf1c85923dba47e841a4",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los roles."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create a Rol",
                "operationId": "4eb1ce389f63b4dbfcc5b002014636e1",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The title of the departments",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Los campos no son validos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Mostrar un rol concreto",
                "operationId": "d932a477623c1c931444dcc799f1eb0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar el rol especificado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Update a Roles",
                "operationId": "e0fa8b9f9f9cfb6c7e51a7609dd35dcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The title of the role",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al procesa la solicitud"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Eliminar un roles concreto",
                "operationId": "a0c87b804be9d37055fc6f936cbacb1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Aceppted."
                    },
                    "204": {
                        "description": "No content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}